Change capacity and amperage on Integer

+++ src/main/java/hibernate/dao/Interface/IBattery.java
@@ -5,7 +5,7 @@ import java.util.Collection;
 public interface IBattery<T> extends BaseCommand<T>{
     T getBatteryById(final int id);
     Collection<T> getByMark(final String mark);
-    Collection<T> getByCapacity(final double capacity);
-    Collection<T> getByAmperage(final double amperage);
+    Collection<T> getByCapacity(final int capacity);
+    Collection<T> getByAmperage(final int amperage);
     Collection<T> getByPolarityId(final int polarityId);
 }
